+++ src/lodepng.c
@@ -2409,7 +2409,7 @@ void lodepng_chunk_type(char type[5], const unsigned char* chunk)
 
 unsigned char lodepng_chunk_type_equals(const unsigned char* chunk, const char* type)
 {
-    if(strlen(type) != 4) return 0;
+    if((int)strlen(type) != 4) return 0;
     return (chunk[4] == type[0] && chunk[5] == type[1] && chunk[6] == type[2] && chunk[7] == type[3]);
 }
 
@@ -5697,12 +5697,12 @@ unsigned lodepng_encode(unsigned char** out, size_t* outsize,
         /*tEXt and/or zTXt*/
         for(i = 0; i != info.text_num; ++i)
         {
-            if(strlen(info.text_keys[i]) > 79)
+            if((int)strlen(info.text_keys[i]) > 79)
             {
                 state->error = 66; /*text chunk too large*/
                 break;
             }
-            if(strlen(info.text_keys[i]) < 1)
+            if((int)strlen(info.text_keys[i]) < 1)
             {
                 state->error = 67; /*text chunk too small*/
                 break;
@@ -5736,12 +5736,12 @@ unsigned lodepng_encode(unsigned char** out, size_t* outsize,
         /*iTXt*/
         for(i = 0; i != info.itext_num; ++i)
         {
-            if(strlen(info.itext_keys[i]) > 79)
+            if((int)strlen(info.itext_keys[i]) > 79)
             {
                 state->error = 66; /*text chunk too large*/
                 break;
             }
-            if(strlen(info.itext_keys[i]) < 1)
+            if((int)strlen(info.itext_keys[i]) < 1)
             {
                 state->error = 67; /*text chunk too small*/
                 break;
